requirements.txt
bctpy
networkx
nibabel
nilearn
nipy
nipype
numpy
pandas
pickleshare
scikit-image
scikit-learn
scipy
seaborn
sklearn
